Create Let's enhance the contracts by adding a basic token transfer functionality along with event logging for better visibility
Added an amount parameter to both sendToB in ContractA and sendToA in ContractB to represent the number of tokens to transfer.

Used starknet.transfer to simulate a token transfer between contracts.

Introduced a Transfer event to log the details of token transfers. This event can be helpful for external systems or user interfaces to track and display relevant information.